Move feature to main server.xml
Features in an XML that is included via a path containing a variable are not installed by features.sh.  So the transportSecurity-1.0 feature that was listed in the included oidc.xml wasn't getting added to my kernel-slim image, meaning it wasn't available within the container image and thus didn't load.  So moving it up to the main server.xml to ensure it gets handled by features.sh